RaspberryPI/services/client-mqtt/app/requirements.txt
paho-mqtt==2.1.0
pyserial==3.5
setuptools==72.1.0
wheel==0.44.0
